+++ src/actions/todolistActions.ts
@@ -1,0 +1,7 @@
+import * as actionTypes from './types/todolistTypes';
+import {ISingleElementList} from '../entities/todoSingleElement'
+
+export const setNewElementToDoList = (newElement:ISingleElementList) =>{(
+    type: actionTypes.SET_NEW_ELEMENT,
+    newElement
+)}