+++ src/components/Form.tsx
@@ -1,0 +1,9 @@
+import React, {FC, useState} from 'react';
+
+const Form: FC = props =>{
+    return{
+
+    }
+};
+
+export default Form;