+++ src/entities/todoSingleElement.ts
@@ -1,0 +1,4 @@
+export interface ISingleElementList{
+   title:string;
+    description: string;
+}