+++ CHANGELOG.md
@@ -1,15 +1,48 @@
 # Changelog
-All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [0.0.5] - 2020-05-04
 
-## [1.0.0] - 2020-02-25
+**Changes**
 
-### Changes
+-   Styled profile screen
 
-    - 
+**Added**
 
-### Added
+-   Entities: todoSingleElement
+-   Reducers: todoListReducer, index
+-   Actions: todoListTypes, todolistActions
+-   Compononents: CreateListButton, ListContainer, Form
+-   Readme: instruction how to run project
 
-    - Initial app config
+## [0.0.4] - 2020-04-18
+
+**Changes**
+
+-   Styled home screen
+-   Redesign profile screen
+
+
+## [0.0.3] - 2020-04-16
+
+**Changes**
+
+-   Styled profile screen
+-   Change splash page icon
+
+**Added**
+
+-   Screens: profile, home, todolist
+-   React-Redux Styled Components
+
+## [0.0.2] - 2020-04-13
+
+**Added**
+
+-   Compononents: BottomNavigation, StatusBar
+-   Basic structure of the project
+
+## [0.0.1] - 2020-04-01
+
+**Added**
+
+-  Initial app config